Fixed missing constants bug
The GUI that comes with JNEAT usually allows you to load in the default
values for 20 constants in the Neat class. I found the file they load
from and set them equal to the correct values in Java code. I should be
on the look out for more error like this...