Achieved a high score of 12,000
After lots of changes to my code, including adding an input scaling
method, the NEAT algorithm seems to slowly evolve controllers into the 8
to 10 thousand range. I believe that this could have been even higher
given longer.

Next I need to finish of all of my stories formally. I need to refactor
and finish the coding and commenting, and carry out formal tests. For my
5th sprint I will need to refactor the executor class and impliment the
final story. This will have to include pausing/ending the program to see
progress, the ability to extend the running of the program, and the
saving of champion genomes every [X] cycles that can be loaded and
played at the users discrecion.