Written command line menu. Refactored to allow multiple NEAT controllers.
Comman line menu allows for the loading of a file by filename and
running it in a sim, or loading it and running a chosen number of games
to find min, max and average scores.

The refactoring allows programmers to extend a controller class
'PacManController' and impliment methods for finding network inputs.
This allows for controllers with different input configurations to be
made.